+++ R/colors.R
@@ -278,10 +278,10 @@ S7::method(str, class_color) <- function(
     nest.lev = nest.lev)
 }
 
-S7::method(print, class_color) <- function(x, ...) {
-  str(x, ...)
-  invisible(x)
-}
+# S7::method(print, class_color) <- function(x, ...) {
+#   str(x, ...)
+#   invisible(x)
+# }
 
 S7::method(`[`, class_color) <- function(x, i) {
   i <- character_index(i, x@id)