+++ R/ggdiagram-package.R
@@ -3,8 +3,8 @@
 
 ## usethis namespace: start
 #' @import S7
-#' @rawNamespace if (packageVersion("ggplot2") >= "3.5.2.9000") importFrom("ggplot2", "update_ggplot")
-#' @rawNamespace if (packageVersion("ggplot2") >= "3.5.2.9000") importFrom("ggplot2", "class_ggplot")
+#' @importFrom ggplot2 update_ggplot
+#' @importFrom ggplot2 class_ggplot
 #' @importFrom rlang .data
 #' @importFrom rlang `%||%`
 ## usethis namespace: end