+++ NEWS.md
@@ -1,3 +1,11 @@
+# ggdiagram 0.1.1 _2025-09-14_
+
+* ggplot2 4.0.0 broke ggdiagram 0.1.0. New version requires ggplot2 4.0.0 or higher. 
+* Fix print method for `class_color`
+* Fix `@tangent_at` for ellipses when given a point
+
+
+
 # ggdiagram 0.1.0 _2025-08-18_
 
 * Initial CRAN submission.